🚧 Update session start logic, fix dim on slide out experiences

+++ Sources/UserPilot/Experiences/Theme/Models/ThemeContent.swift
@@ -145,7 +145,7 @@ internal struct ExperienceTheme: Codable {
     }
 
     var backdropOpacity: CGFloat {
-        CGFloat((backdrop?.opacity ?? ThemeHandler.DefaultValues.dimDegree / 100))
+        CGFloat(CGFloat(backdrop?.opacity ?? ThemeHandler.DefaultValues.dimDegree) / 100)
     }
 
     var backdropBackground: UIColor {
